UI/frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=4.3.0",
    "pyyaml>=6.0.2",
    "streamlit>=1.47.0",
    "websockets>=15.0.1",
]
